Reorganise StepItem and sub-components
