+++ app/pages/lab-pages-editor/components/TasksPage/components/StepItem/StepItem.jsx
@@ -5,17 +5,18 @@
 /* eslint-disable react/jsx-boolean-value */
 /* eslint-disable react/forbid-prop-types */
 
-import { useState } from 'react';
 import PropTypes from 'prop-types';
 
 // import strings from '../../../strings.json'; // TODO: move all text into strings
+import DropTarget from './DropTarget.jsx';
 import TaskItem from './TaskItem.jsx';
-import CopyIcon from '../../../icons/CopyIcon.jsx';
-import DeleteIcon from '../../../icons/DeleteIcon.jsx';
-import EditIcon from '../../../icons/EditIcon.jsx';
-import GripIcon from '../../../icons/GripIcon.jsx';
-import MoveDownIcon from '../../../icons/MoveDownIcon.jsx';
-import MoveUpIcon from '../../../icons/MoveUpIcon.jsx';
+
+import CopyIcon from '../../../../icons/CopyIcon.jsx';
+import DeleteIcon from '../../../../icons/DeleteIcon.jsx';
+import EditIcon from '../../../../icons/EditIcon.jsx';
+import GripIcon from '../../../../icons/GripIcon.jsx';
+import MoveDownIcon from '../../../../icons/MoveDownIcon.jsx';
+import MoveUpIcon from '../../../../icons/MoveUpIcon.jsx';
 
 function StepItem({
   activeDragItem = -1,
@@ -141,56 +142,5 @@ StepItem.propTypes = {
   stepKey: PropTypes.string,
   stepIndex: PropTypes.number
 };
-
-function DropTarget({
-  activeDragItem = -1,
-  moveStep = () => {},
-  setActiveDragItem = () => {},
-  targetIndex = 0
-}) {
-  const [active, setActive] = useState(false);
-
-  function onDragEnter(e) {
-    const from = activeDragItem;
-    const to = (from < targetIndex) ? targetIndex - 1 : targetIndex; 
-    setActive(from !== to);
-    e.preventDefault(); // Prevent default, to ensure onDrop works.
-  }
-
-  function onDragLeave(e) {
-    setActive(false);
-    e.preventDefault();  // Probably unnecessary for onDrop, but oh well
-  }
-
-  function onDragOver(e) {
-    e.preventDefault(); // Prevent default, to ensure onDrop works.
-  }
-
-  function onDrop(e) {
-    const from = parseInt(e.dataTransfer.getData('text/plain')) || 0;
-    const to = (from < targetIndex) ? targetIndex - 1 : targetIndex;
-    moveStep(from, to);
-    setActive(false);
-    setActiveDragItem(-1);
-    e.preventDefault();
-  }
-
-  return (
-    <div
-      className={`step-drop-target ${active ? 'active' : ''}`}
-      onDragEnter={onDragEnter}
-      onDragLeave={onDragLeave}
-      onDragOver={onDragOver}
-      onDrop={onDrop}
-    ></div>
-  );
-}
-
-DropTarget.propTypes = {
-  activeDragItem: PropTypes.number,
-  moveStep: PropTypes.func,
-  setActiveDragItem: PropTypes.func,
-  targetIndex: PropTypes.number
-}
 
 export default StepItem;
