+++ app/pages/lab-pages-editor/components/TasksPage/components/StepItem/index.jsx
@@ -1,0 +1,3 @@
+import StepItem from './StepItem.jsx'
+
+export default StepItem;
